upgrade armanino to v3
