+++ .changeset/thin-jobs-push.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/armanino-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
