+++ .pnp.cjs
@@ -228,10 +228,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         "reference": "workspace:packages/sources/armanino"\
       },\
       {\
-        "name": "@chainlink/armanino-test-adapter",\
-        "reference": "workspace:packages/sources/armanino-test"\
-      },\
-      {\
         "name": "@chainlink/avalanche-platform-adapter",\
         "reference": "workspace:packages/sources/avalanche-platform"\
       },\
@@ -834,7 +830,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
       ["@chainlink/apy-finance-adapter", ["workspace:packages/composites/apy-finance"]],\
       ["@chainlink/apy-finance-test-adapter", ["workspace:packages/composites/apy-finance-test"]],\
       ["@chainlink/armanino-adapter", ["workspace:packages/sources/armanino"]],\
-      ["@chainlink/armanino-test-adapter", ["workspace:packages/sources/armanino-test"]],\
       ["@chainlink/augur-adapter", ["workspace:packages/composites/augur"]],\
       ["@chainlink/avalanche-platform-adapter", ["workspace:packages/sources/avalanche-platform"]],\
       ["@chainlink/bank-frick-adapter", ["workspace:packages/sources/bank-frick"]],\
@@ -5726,25 +5721,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "packageLocation": "./packages/sources/armanino/",\
           "packageDependencies": [\
             ["@chainlink/armanino-adapter", "workspace:packages/sources/armanino"],\
-            ["@chainlink/ea-bootstrap", "workspace:packages/core/bootstrap"],\
-            ["@chainlink/ea-test-helpers", "workspace:packages/core/test-helpers"],\
-            ["@types/jest", "npm:27.5.2"],\
-            ["@types/node", "npm:16.11.51"],\
-            ["@types/supertest", "npm:2.0.12"],\
-            ["nock", "npm:13.2.9"],\
-            ["supertest", "npm:6.2.4"],\
-            ["tslib", "npm:2.4.1"],\
-            ["typescript", "patch:typescript@npm%3A5.0.4#~builtin<compat/typescript>::version=5.0.4&hash=b5f058"]\
-          ],\
-          "linkType": "SOFT"\
-        }]\
-      ]],\
-      ["@chainlink/armanino-test-adapter", [\
-        ["workspace:packages/sources/armanino-test", {\
-          "packageLocation": "./packages/sources/armanino-test/",\
-          "packageDependencies": [\
-            ["@chainlink/armanino-test-adapter", "workspace:packages/sources/armanino-test"],\
-            ["@chainlink/external-adapter-framework", "npm:0.30.2"],\
+            ["@chainlink/external-adapter-framework", "npm:0.31.0"],\
             ["@types/jest", "npm:27.5.2"],\
             ["@types/node", "npm:16.11.51"],\
             ["nock", "npm:13.2.9"],\
