--- packages/sources/armanino-test/src/endpoint/index.ts
@@ -1,2 +1,0 @@
-export { endpoint as mco2 } from './mco2'
-export { endpoint as stbt } from './stbt'
