--- packages/sources/armanino-test/src/endpoint/mco2.ts
@@ -1,17 +1,0 @@
-import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
-import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
-import { config } from '../config'
-import { httpTransport } from '../transport/mco2'
-import { EmptyInputParameters } from '@chainlink/external-adapter-framework/validation/input-params'
-
-export type BaseEndpointTypes = {
-  Parameters: EmptyInputParameters
-  Response: SingleNumberResultResponse
-  Settings: typeof config.settings
-}
-
-export const endpoint = new AdapterEndpoint({
-  name: 'mco2',
-  aliases: ['balance'],
-  transport: httpTransport,
-})
