--- packages/sources/armanino/schemas/env.json
@@ -1,19 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/armanino-adapter.json",
-  "title": "Armanino Adapter",
-  "description": "Carbon credit reserves attested by Armanino",
-  "required": [],
-  "type": "object",
-  "properties": {
-    "API_ENDPOINT": {
-      "default": "https://api.oracle-services.ledgerlens.io/v1/chainlink/proof-of-reserves/",
-      "description": "API Endpoint to use",
-      "type": "string"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
