--- packages/sources/armanino/src/config/limits.json
@@ -1,1 +1,0 @@
-{ "http": {}, "ws": {} }
