--- packages/sources/armanino/src/config/symbols.json
@@ -1,1 +1,0 @@
-{}
