+++ packages/sources/armanino/src/endpoint/index.ts
@@ -1,6 +1,2 @@
-import type { TInputParameters as MCO2InputParameters } from './mco2'
-
-export type TInputParameters = MCO2InputParameters
-
-export * as mco2 from './mco2'
-export * as stbt from './stbt'
+export { endpoint as mco2 } from './mco2'
+export { endpoint as stbt } from './stbt'
