--- packages/sources/armanino/test/e2e/adapter.test.ts
@@ -1,55 +1,0 @@
-import type { AdapterRequest, FastifyInstance } from '@chainlink/ea-bootstrap'
-import request, { SuperTest, Test } from 'supertest'
-import * as process from 'process'
-import { server as startServer } from '../../src'
-import * as nock from 'nock'
-import { AddressInfo } from 'net'
-import { setEnvVariables } from '@chainlink/ea-test-helpers'
-
-describe('execute', () => {
-  let fastify: FastifyInstance
-  let req: SuperTest<Test>
-  let oldEnv: NodeJS.ProcessEnv
-
-  beforeAll(async () => {
-    oldEnv = JSON.parse(JSON.stringify(process.env))
-    process.env.CACHE_ENABLED = 'false'
-
-    if (process.env.RECORD) {
-      nock.recorder.rec()
-    }
-
-    fastify = await startServer()
-    req = request(`localhost:${(fastify.server.address() as AddressInfo).port}`)
-  })
-  afterAll((done) => {
-    setEnvVariables(oldEnv)
-
-    if (process.env.RECORD) {
-      nock.recorder.play()
-    }
-
-    nock.restore()
-    nock.cleanAll()
-    nock.enableNetConnect()
-    fastify.close(done)
-  })
-
-  describe('mco2 endpoint', () => {
-    const balanceRequest: AdapterRequest = {
-      id: '1',
-      data: {},
-    }
-
-    it('should return success', async () => {
-      const response = await req
-        .post('/')
-        .send(balanceRequest)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-})
