+++ packages/sources/armanino/test/integration/adapter.test.ts
@@ -1,72 +1,52 @@
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import { server as startServer } from '../../src'
-import { mockMCO2Response, mockSTBTResponseSuccess, mockSTBTResponseFailure } from './fixtures'
-import { setupExternalAdapterTest } from '@chainlink/ea-test-helpers'
-import type { SuiteContext } from '@chainlink/ea-test-helpers'
-import { SuperTest, Test } from 'supertest'
+import {
+  TestAdapter,
+  setEnvVariables,
+} from '@chainlink/external-adapter-framework/util/testing-utils'
+import * as nock from 'nock'
+import { mockMCO2Response, mockSTBTResponse } from './fixtures'
 
 describe('execute', () => {
-  const context: SuiteContext = {
-    req: null,
-    server: startServer,
-  }
+  let spy: jest.SpyInstance
+  let testAdapter: TestAdapter
+  let oldEnv: NodeJS.ProcessEnv
 
-  const envVariables = {
-    CACHE_ENABLED: 'false',
-  }
+  beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
+    const mockDate = new Date('2001-01-01T11:11:11.111Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
 
-  setupExternalAdapterTest(envVariables, context)
-
-  describe('mco2 endpoint', () => {
-    const request: AdapterRequest = {
-      id: '1',
-      data: {},
-    }
-
-    it('should return success', async () => {
-      mockMCO2Response()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(request)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+    const adapter = (await import('./../../src')).adapter
+    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
+      testAdapter: {} as TestAdapter<never>,
     })
   })
-  describe('stbt endpoint', () => {
-    const request: AdapterRequest = {
-      id: '1',
-      data: {
-        endpoint: 'stbt',
-      },
-    }
 
+  afterAll(async () => {
+    setEnvVariables(oldEnv)
+    await testAdapter.api.close()
+    nock.restore()
+    nock.cleanAll()
+    spy.mockRestore()
+  })
+
+  describe('mco2 endpoint', () => {
     it('should return success', async () => {
-      mockSTBTResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(request)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      mockMCO2Response()
+      const response = await testAdapter.request({})
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
+  })
 
-    it('should return error when ripcord true', async () => {
-      mockSTBTResponseFailure()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(request)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-      expect(response.body).toMatchSnapshot()
+  describe('stbt endpoint', () => {
+    it('should return success', async () => {
+      const data = {
+        endpoint: 'stbt',
+      }
+      mockSTBTResponse()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 })
